session-2/requirements.txt
matplotlib
numpy
Pillow
torch
ray
torchvision
pandas